add security check

+++ app/controllers/delayed_jobs_controller.rb
@@ -6,6 +6,8 @@ class DelayedJobsController < ApplicationController
   end
 
   def destroy
+    authorize @job
+
     if @job.destroy
       flash[:notice] = "Job was successfully deleted."
     else
